+++ Com.DanLiris.Service.Purchasing.Lib/Interfaces/IGarmentInvoice.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 
 namespace Com.DanLiris.Service.Purchasing.Lib.Interfaces
 {
-    public interface IGarmentInvoice 
+    public interface IGarmentInvoice
     {
 		Tuple<List<GarmentInvoice>, int, Dictionary<string, string>> Read(int Page = 1, int Size = 25, string Order = "{}", string Keyword = null, string Filter = "{}");
 		GarmentInvoice ReadById(int id);
 		Task<int> Create(GarmentInvoice m, string user, int clientTimeZoneOffset = 7);
 		Task<int> Update(int id, GarmentInvoice m, string user, int clientTimeZoneOffset = 7);
 		int Delete(int id, string username);
-        GarmentInvoice ReadByDOId(long id);
+        GarmentInvoice ReadByDOId(int id);
 	}
 }
